requirements.txt
PyPDF2
Pillow
pandas
streamlit
re